refactor: move odos API queries to utils
