+++ scripts/periphery/leverage_zap.py
@@ -1,12 +1,7 @@
-import requests
-from web3.main import to_checksum_address
-
 from brownie import Contract, chain
 from scripts.utils.float2int import to_int
+from .utils import odos
 
-# https://docs.odos.xyz/api/endpoints/
-QUOTE_URL = "https://api.odos.xyz/sor/quote/v2"
-ASSEMBLE_URL = "https://api.odos.xyz/sor/assemble"
 
 CONTROLLER = "0x1337F001E280420EcCe9E7B934Fa07D67fdb62CD"
 MONEY = "0x69420f9e38a4e60a62224c489be4bf7a94402496"
@@ -65,15 +60,15 @@ def get_close_loan_routing_data(zap, account, market, use_account_balance=True, 
     assert shortfall > 0
 
     amount_out = shortfall / controller.get_oracle_price(token) * (1 + max_slippage)
-    amount_out = to_int(amount, token.decimals())
+    amount_out = to_int(amount_out, token.decimals())
 
-    amount_in, path_id = get_quote(chain.id, zap, token, MONEY, amount_out, max_slippage)
+    amount_in, path_id = odos.get_quote(chain.id, zap, token, MONEY, amount_out, max_slippage)
 
     while amount_in < shortfall:
         amount_out = int(amount_out * shortfall / amount_in)
-        amount_in, path_id = get_quote(chain.id, zap, token, MONEY, amount_out, max_slippage)
+        amount_in, path_id = odos.get_quote(chain.id, zap, token, MONEY, amount_out, max_slippage)
 
-    return get_route_calldata(zap, path_id), amount_out, amount_in
+    return odos.get_route_calldata(zap, path_id), amount_out, amount_in
 
 
 def get_add_coll_routing_data(zap, account, market, coll_amount, max_slippage=0.003):
@@ -103,60 +98,6 @@ def get_add_coll_routing_data(zap, account, market, coll_amount, max_slippage=0.
     debt_amm, coll_amm = Contract(amm).get_sum_xy(account)
     assert debt_amm > 0
 
-    received, path_id = get_quote(chain.id, zap, MONEY, token, debt_amm, max_slippage)
+    received, path_id = odos.get_quote(chain.id, zap, MONEY, token, debt_amm, max_slippage)
 
-    return get_route_calldata(zap, path_id), coll_amount + coll_amm + received
-
-
-def get_quote(chain_id, caller, input_token, output_token, amount_in, max_slippage=0.003):
-    """
-    Args:
-        amount_in: Amount of input_token being swapped, as an integer with the same precision
-            used in the token smart contract
-    """
-    quote_request_body = {
-        "chainId": chain_id,
-        "inputTokens": [
-            {"tokenAddress": to_checksum_address(str(input_token)), "amount": str(amount_in)}
-        ],
-        "outputTokens": [{"tokenAddress": to_checksum_address(str(output_token)), "proportion": 1}],
-        "slippageLimitPercent": max_slippage * 100,
-        "userAddr": to_checksum_address(str(caller)),
-        "referralCode": 0,
-        "disableRFQs": True,
-        "compact": True,
-    }
-
-    response = requests.post(
-        QUOTE_URL,
-        headers={"Content-Type": "application/json"},
-        json=quote_request_body,
-    )
-
-    if response.status_code != 200:
-        raise ValueError(f"{response.status_code} error during quote: {response.json()}")
-
-    quote = response.json()
-
-    return int(quote["outAmounts"][0]), quote["pathId"]
-
-
-def get_route_calldata(caller, path_id):
-    assemble_request_body = {
-        "userAddr": to_checksum_address(str(caller)),
-        "pathId": path_id,
-        "simulate": False,
-    }
-
-    response = requests.post(
-        ASSEMBLE_URL,
-        headers={"Content-Type": "application/json"},
-        json=assemble_request_body,
-    )
-
-    if response.status_code != 200:
-        raise ValueError(f"{response.status_code} error during assembly: {response.json()}")
-
-    data = response.json()
-
-    return data["transaction"]["data"]
+    return odos.get_route_calldata(zap, path_id), coll_amount + coll_amm + received
